Fixes implicit narrowing conversion reported in #9166.

+++ jetty-core/jetty-server/src/main/java/org/eclipse/jetty/server/internal/HttpConnection.java
@@ -828,7 +828,7 @@ public class HttpConnection extends AbstractConnection implements Runnable, Writ
                             BufferUtil.clear(_content);
                         }
 
-                        byte gatherWrite = 0;
+                        int gatherWrite = 0;
                         long bytes = 0;
                         if (BufferUtil.hasContent(_header))
                         {
